Shield BigDecimal from SWADL

+++ src/main/java/com/symphony/devsol/federation/activity/AddConnectRoomMember.java
@@ -1,16 +1,17 @@
 package com.symphony.devsol.federation.activity;
 
 import com.symphony.bdk.workflow.swadl.v1.activity.BaseActivity;
+
 import lombok.Data;
 import lombok.EqualsAndHashCode;
-import java.math.BigDecimal;
+
 import java.util.List;
 
 @Data
 @EqualsAndHashCode(callSuper=true)
 public class AddConnectRoomMember extends BaseActivity {
     private String externalNetwork;
-    private BigDecimal advisorSymphonyId;
-    private List<BigDecimal> memberSymphonyIds;
+    private long advisorSymphonyId;
+    private List<Long> memberSymphonyIds;
     private List<String> streamIds;
 }
